remove nunjucks abstractions
